requirements.txt
opencv-python
Pillow
pyautogui
numpy